Move Storage files to the expected path
