+++ lib/fog/parsers/storage/google/access_control_list.rb
