+++ lib/fog/parsers/storage/google/copy_object.rb
