+++ lib/fog/parsers/storage/google/get_bucket.rb
