+++ lib/fog/parsers/storage/google/get_bucket_logging.rb
