+++ lib/fog/parsers/storage/google/get_request_payment.rb
