+++ lib/fog/parsers/storage/google/get_service.rb
