+++ lib/fog/storage/google_json/models/directory.rb
@@ -1,16 +1,11 @@
-require "fog/core/model"
-require "fog/google/models/storage/files"
-
 module Fog
   module Storage
-    class GoogleXML
+    class GoogleJSON
       class Directory < Fog::Model
         identity :key, :aliases => %w(Name name)
 
-        attribute :creation_date, :aliases => "CreationDate"
-
         def acl=(new_acl)
-          valid_acls = ["private", "public-read", "public-read-write", "authenticated-read"]
+          valid_acls = %w(private projectPrivate publicRead publicReadWrite authenticatedRead)
           unless valid_acls.include?(new_acl)
             raise ArgumentError.new("acl must be one of [#{valid_acls.join(', ')}]")
           end
@@ -27,7 +22,7 @@ module Fog
 
         def files
           @files ||= begin
-            Fog::Storage::GoogleXML::Files.new(
+            Fog::Storage::GoogleJSON::Files.new(
               :directory => self,
               :service => service
             )
@@ -36,7 +31,7 @@ module Fog
 
         def public=(new_public)
           if new_public
-            @acl = "public-read"
+            @acl = "publicRead"
           else
             @acl = "private"
           end
@@ -45,7 +40,8 @@ module Fog
 
         def public_url
           requires :key
-          if service.get_bucket_acl(key).body["AccessControlList"].detect { |entry| entry["Scope"]["type"] == "AllUsers" && entry["Permission"] == "READ" }
+          acl = service.get_bucket_acl(key).body
+          if acl["items"].detect { |entry| entry["entity"] == "allUsers" && entry["role"] == "READER" }
             if key.to_s =~ /^(?:[a-z]|\d(?!\d{0,2}(?:\.\d{1,3}){3}$))(?:[a-z0-9]|\.(?![\.\-])|\-(?![\.])){1,61}[a-z0-9]$/
               "https://#{key}.storage.googleapis.com"
             else
@@ -57,7 +53,7 @@ module Fog
         def save
           requires :key
           options = {}
-          options["x-goog-acl"] = @acl if @acl
+          options["predefinedAcl"] = @acl if @acl
           options["LocationConstraint"] = @location if @location
           options["StorageClass"] = attributes[:storage_class] if attributes[:storage_class]
           service.put_bucket(key, options)
