+++ lib/fog/storage/google_json/models/file.rb
@@ -1,5 +1,3 @@
-require "fog/core/model"
-
 module Fog
   module Storage
     class GoogleJSON
