+++ lib/fog/storage/google_json/requests/delete_bucket.rb
