+++ lib/fog/storage/google_json/requests/get_object_url.rb
