+++ lib/fog/storage/google_json/requests/get_service.rb
@@ -2,8 +2,6 @@ module Fog
   module Storage
     class GoogleJSON
       class Real
-        require "fog/google/parsers/storage/get_service"
-
         # TODO: Implement in JSON API
 
         # List information about Google Storage buckets for authorized user
