+++ lib/fog/storage/google_json/requests/head_object.rb
