+++ lib/fog/storage/google_json/requests/put_bucket.rb
