+++ lib/fog/storage/google_json/requests/put_object_url.rb
