+++ lib/fog/storage/google_xml.rb
@@ -4,13 +4,13 @@ module Fog
       requires :google_storage_access_key_id, :google_storage_secret_access_key
       recognizes :host, :port, :scheme, :persistent, :path_style
 
-      model_path "fog/google/models/storage"
+      model_path "fog/storage/google_xml/models"
       collection :directories
       model :directory
       collection :files
       model :file
 
-      request_path "fog/google/requests/storage"
+      request_path "fog/storage/google_xml/requests"
       request :copy_object
       request :delete_bucket
       request :delete_object
