+++ lib/fog/storage/google_xml/models/directory.rb
@@ -1,14 +1,13 @@
-require "fog/core/model"
-require "fog/google/models/storage_json/files"
-
 module Fog
   module Storage
-    class GoogleJSON
+    class GoogleXML
       class Directory < Fog::Model
         identity :key, :aliases => %w(Name name)
 
+        attribute :creation_date, :aliases => "CreationDate"
+
         def acl=(new_acl)
-          valid_acls = %w(private projectPrivate publicRead publicReadWrite authenticatedRead)
+          valid_acls = ["private", "public-read", "public-read-write", "authenticated-read"]
           unless valid_acls.include?(new_acl)
             raise ArgumentError.new("acl must be one of [#{valid_acls.join(', ')}]")
           end
@@ -25,7 +24,7 @@ module Fog
 
         def files
           @files ||= begin
-            Fog::Storage::GoogleJSON::Files.new(
+            Fog::Storage::GoogleXML::Files.new(
               :directory => self,
               :service => service
             )
@@ -34,7 +33,7 @@ module Fog
 
         def public=(new_public)
           if new_public
-            @acl = "publicRead"
+            @acl = "public-read"
           else
             @acl = "private"
           end
@@ -43,8 +42,7 @@ module Fog
 
         def public_url
           requires :key
-          acl = service.get_bucket_acl(key).body
-          if acl["items"].detect { |entry| entry["entity"] == "allUsers" && entry["role"] == "READER" }
+          if service.get_bucket_acl(key).body["AccessControlList"].detect { |entry| entry["Scope"]["type"] == "AllUsers" && entry["Permission"] == "READ" }
             if key.to_s =~ /^(?:[a-z]|\d(?!\d{0,2}(?:\.\d{1,3}){3}$))(?:[a-z0-9]|\.(?![\.\-])|\-(?![\.])){1,61}[a-z0-9]$/
               "https://#{key}.storage.googleapis.com"
             else
@@ -56,7 +54,7 @@ module Fog
         def save
           requires :key
           options = {}
-          options["predefinedAcl"] = @acl if @acl
+          options["x-goog-acl"] = @acl if @acl
           options["LocationConstraint"] = @location if @location
           options["StorageClass"] = attributes[:storage_class] if attributes[:storage_class]
           service.put_bucket(key, options)
