+++ lib/fog/storage/google_xml/requests/delete_bucket.rb
