+++ lib/fog/storage/google_xml/requests/delete_object.rb
