+++ lib/fog/storage/google_xml/requests/get_bucket_acl.rb
@@ -2,15 +2,10 @@ module Fog
   module Storage
     class GoogleXML
       class Real
-        require "fog/google/parsers/storage/access_control_list"
-
-        # Get access control list for an Google Storage object
+        # Get access control list for an Google Storage bucket
         #
         # ==== Parameters
-        # * bucket_name<~String> - name of bucket containing object
-        # * object_name<~String> - name of object to get access control list for
-        # * options<~Hash>:
-        #   * 'versionId'<~String> - specify a particular version to retrieve
+        # * bucket_name<~String> - name of bucket to get access control list for
         #
         # ==== Returns
         # * response<~Excon::Response>:
@@ -28,28 +23,22 @@ module Fog
         #              * 'URI'<~String> - URI of group to grant access for
         #           * 'Permission'<~String> - Permission, in [FULL_CONTROL, WRITE, WRITE_ACP, READ, READ_ACP]
         #
-        def get_object_acl(bucket_name, object_name, options = {})
+        def get_bucket_acl(bucket_name)
           raise ArgumentError.new("bucket_name is required") unless bucket_name
-          raise ArgumentError.new("object_name is required") unless object_name
-          query = { "acl" => nil }
-          if version_id = options.delete("versionId")
-            query["versionId"] = version_id
-          end
           request(:expects    => 200,
                   :headers    => {},
                   :host       => "#{bucket_name}.#{@host}",
                   :idempotent => true,
                   :method     => "GET",
                   :parser     => Fog::Parsers::Storage::Google::AccessControlList.new,
-                  :path       => CGI.escape(object_name),
-                  :query      => query)
+                  :query      => { "acl" => nil })
         end
       end
 
       class Mock
-        def get_object_acl(bucket_name, object_name)
+        def get_bucket_acl(bucket_name)
           response = Excon::Response.new
-          if acl = data[:acls][:object][bucket_name] && data[:acls][:object][bucket_name][object_name]
+          if acl = data[:acls][:bucket][bucket_name]
             response.status = 200
             response.body = acl
           else
