+++ lib/fog/storage/google_xml/requests/get_object_http_url.rb
