+++ lib/fog/storage/google_xml/requests/head_object.rb
