+++ lib/fog/storage/google_xml/requests/put_bucket.rb
