+++ lib/fog/storage/google_xml/requests/put_object_url.rb
